Filter lesser and greater episodes
